Refactored dsl query objects into new package
