+++ search/dsl/rescore_query.py
@@ -1,0 +1,8 @@
+"""
+Defines the DSL for a rescore query
+"""
+from elasticsearch_dsl import query as Q
+
+
+class RescoreQuery(Q.Query):
+    name = "rescore"
